# documentation/swagger.yaml
openapi: 3.0.3
info:
  title: Learn Wise API App
  description: |-
    lorem ipsum dolor is amet
  contact:
    email: mahadiocsidio@students.amikom.ac.id
  version: 1.0.0
servers:
  - url: https://binar-academy-final-project.vercel.app
...
paths:
...
  /auth/verify-otp:
    post:
      tags:
        - Start
      operationId: verify-otp
      summary: Verify OTP
      requestBody:
        description: Verify OTP request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                otp:
                  type: string
                  example: 123456
      responses:
        '200':
          description: OTP verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Activation Code verified successfully'
                  err:
                    type: string
                    example: null
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                        example: user@example.com
                      otp:
                        type: string
                        example: 123456
                      is_verified:
                        type: boolean
                        example: true
        '400':
          description: Failed OTP verification
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Bad Request'
                  error:
                    type: string
                    example: 'Invalid Email or OTP'
                  data:
                    example: null
...
  /auth/reset-password:
    post:
      tags:
        - Start
      operationId: reset-password
      summary: Reset Password
      requestBody:
        description: Reset password request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Password reset email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Send'
                  err:
                    type: string
                    example: null
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                        example: user@example.com
        '400':
          description: Email not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Email Not Found'
                  err:
                    type: string
                    example: 'Enter Registered Email!'
                  data:
                    example: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...